requirements.txt
langchain>=0.0.267
openai>=0.27.8
chromadb>=0.4.13
python-dotenv>=1.0.0
langchain-openai>=0.0.2
tiktoken>=0.5.1
pypdf>=3.15.1
